pyproject.toml
[tool.poetry]
name = "dns-automation"
version = "0.1.0"
description = ""
authors = ["sajad <fallahdoostsajad@gmal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
requests = "^2.32.3"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
